requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 25.7.0
aiohappyeyeballs=2.4.0=pypi_0
aiohttp=3.10.5=pypi_0
aiosignal=1.3.1=pypi_0
alembic=1.15.2=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.4.0=pypi_0
appnope=0.1.4=pyhd8ed1ab_1
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asteroid-filterbanks=0.4.0=pypi_0
asttokens=3.0.0=pyhd8ed1ab_1
async-lru=2.0.4=pypi_0
async-timeout=4.0.3=pypi_0
attrs=24.2.0=pypi_0
babel=2.16.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
blas=1.0=openblas
bleach=6.1.0=pypi_0
brotlicffi=1.0.9.2=py310h313beb8_1
bzip2=1.0.8=h80987f9_6
ca-certificates=2025.8.3=hbd8a1cb_0
certifi=2024.8.30=pypi_0
cffi=1.17.1=py310h3eb5a62_1
charset-normalizer=3.4.3=pypi_0
click=8.1.8=pypi_0
colorlog=6.9.0=pypi_0
comm=0.2.3=pyhe01879c_0
contourpy=1.3.0=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.8.16=py310h0962b89_0
decorator=5.2.1=pyhd8ed1ab_0
defusedxml=0.7.1=pypi_0
docopt=0.6.2=pypi_0
einops=0.8.1=pypi_0
exceptiongroup=1.3.0=pyhd8ed1ab_0
executing=2.2.1=pyhd8ed1ab_0
expat=2.7.1=h313beb8_0
fastjsonschema=2.20.0=pypi_0
filelock=3.16.0=pypi_0
filterpy=1.4.5=pypi_0
fonttools=4.53.1=pypi_0
fqdn=1.5.1=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.10.0=pypi_0
gmp=6.3.0=h313beb8_0
gmpy2=2.2.1=py310h5c1b81f_0
grafog=0.1=pypi_0
graphlime=1.2.0=pypi_0
h11=0.14.0=pypi_0
hf-xet=1.1.0=pypi_0
httpcore=1.0.5=pypi_0
httpx=0.27.2=pypi_0
huggingface-hub=0.31.1=pypi_0
hyperpyyaml=1.2.2=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.7.0=pyhe01879c_1
ipykernel=6.29.4=pypi_0
ipython=8.37.0=pyh8f84b5b_0
ipywidgets=8.1.5=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
json5=0.9.25=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
julius=0.2.7=pypi_0
jupyter-events=0.10.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.14.2=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.8.1=pyh31011fe_0
jupyterlab=4.2.5=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.3=pypi_0
jupyterlab-widgets=3.0.13=pypi_0
kiwisolver=1.4.7=pypi_0
kneed=0.8.5=pypi_0
krb5=1.21.3=h237132a_0
libabseil=20240116.2=cxx17_h313beb8_0
libcxx=20.1.8=h8869778_0
libedit=3.1.20250104=pl5321hafb1f1b_0
libffi=3.4.4=hca03da5_1
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libopenblas=0.3.29=hea593b9_0
libprotobuf=4.25.3=h514c7bf_0
libsodium=1.0.20=h99b78c6_0
lightning=2.5.1.post0=pypi_0
lightning-utilities=0.14.3=pypi_0
llvm-openmp=20.1.8=he822017_0
llvmlite=0.44.0=pypi_0
mako=1.3.10=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.8.3=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mdurl=0.1.2=pypi_0
mistune=3.0.2=pypi_0
more-itertools=10.7.0=pypi_0
mpc=1.3.1=h80987f9_0
mpfr=4.2.1=h80987f9_0
mpmath=1.3.0=py310hca03da5_0
multidict=6.6.4=pypi_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.4=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.5=hee39554_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.3=pypi_0
notebook=7.2.2=pypi_0
notebook-shim=0.2.4=pypi_0
numba=0.61.2=pypi_0
numpy=1.26.4=py310h901140f_1
numpy-base=1.26.4=py310hae06d03_1
omegaconf=2.3.0=pypi_0
openai-whisper=20240930=pypi_0
openssl=3.5.2=he92f556_0
opt-einsum=3.4.0=pypi_0
optuna=4.3.0=pypi_0
overrides=7.7.0=pypi_0
packaging=24.2=pypi_0
pandas=2.2.1=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.5=pyhcf101f3_0
patsy=1.0.1=pypi_0
pexpect=4.9.0=pyhd8ed1ab_1
pgmpy=1.0.0=pypi_0
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=10.4.0=pypi_0
pip=25.2=pyhc872135_0
platformdirs=4.4.0=pyhcf101f3_0
primepy=1.3=pypi_0
prometheus-client=0.20.0=pypi_0
prompt-toolkit=3.0.52=pyha770c72_0
propcache=0.3.1=py310h80987f9_0
protobuf=6.30.2=pypi_0
ps=0.1.0=pypi_0
psutil=7.0.0=py310h254cc4a_0
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pyannote-audio=3.3.2=pypi_0
pyannote-core=5.0.0=pypi_0
pyannote-database=5.1.3=pypi_0
pyannote-metrics=3.2.1=pypi_0
pyannote-pipeline=3.0.1=pypi_0
pycparser=2.22=pypi_0
pyg=2.6.1=py310_torch_2.3.0_cpu
pygments=2.19.2=pyhd8ed1ab_0
pyparsing=3.1.4=pypi_0
pyro-api=0.1.2=pypi_0
pyro-ppl=1.9.1=pypi_0
pysocks=1.7.1=py310hca03da5_0
python=3.10.18=h19e8193_0
python-dateutil=2.9.0.post0=pyhe01879c_2
python-json-logger=2.0.7=pypi_0
pytorch=2.3.1=gpu_mps_py310h7b7e308_100
pytorch-lightning=2.5.1.post0=pypi_0
pytorch-metric-learning=2.8.1=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=py310h80987f9_0
pyzmq=26.2.0=py310h313beb8_0
readline=8.3=h0b18652_0
referencing=0.35.1=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rich=14.0.0=pypi_0
rpds-py=0.20.0=pypi_0
ruamel-yaml=0.18.10=pypi_0
ruamel-yaml-clib=0.2.12=pypi_0
scikit-learn=1.5.2=pypi_0
scipy=1.14.1=pypi_0
seaborn=0.13.2=pypi_0
semver=3.0.4=pypi_0
send2trash=1.8.3=pypi_0
sentencepiece=0.2.0=pypi_0
setuptools=73.0.1=pypi_0
shellingham=1.5.4=pypi_0
six=1.17.0=pyhe01879c_1
sleef=3.5.1=h80987f9_2
sniffio=1.3.1=pypi_0
sortedcontainers=2.4.0=pypi_0
soundfile=0.13.1=pypi_0
soupsieve=2.6=pypi_0
speechbrain=1.0.3=pypi_0
sqlalchemy=2.0.40=pypi_0
sqlite=3.50.2=h79febb2_1
stack_data=0.6.3=pyhd8ed1ab_1
statsmodels=0.14.4=pypi_0
sympy=1.13.1=pypi_0
tabulate=0.9.0=pypi_0
tensorboardx=2.6.2.2=pypi_0
terminado=0.18.1=pypi_0
threadpoolctl=3.5.0=pypi_0
tiktoken=0.9.0=pypi_0
tinycss2=1.3.0=pypi_0
tk=8.6.15=hcd8a7d5_0
tomli=2.2.1=pypi_0
torch-audiomentations=0.12.0=pypi_0
torch-pitch-shift=1.2.5=pypi_0
torchaudio=2.6.0=pypi_0
torchmetrics=1.8.2=pypi_0
tornado=6.5.1=py310h80987f9_0
tqdm=4.66.5=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
typer=0.15.3=pypi_0
types-python-dateutil=2.9.0.20240906=pypi_0
typing_extensions=4.15.0=pyhcf101f3_0
tzdata=2024.1=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.2.3=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_1
webcolors=24.8.0=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
wheel=0.44.0=pypi_0
widgetsnbextension=4.0.13=pypi_0
xz=5.6.4=h80987f9_1
yaml=0.2.5=h1a28f6b_0
yarl=1.11.1=pypi_0
zeromq=4.3.5=h888dc83_9
zipp=3.23.0=pyhd8ed1ab_0
zlib=1.2.13=ha1e9f54_2
